added new functions about obstacle_detection part. Those are include:
-checkingObstacleProximity

this function after obstacledetected we are checking the closest 2  node and also adding those nodes as obstacle for safety zone.

-updateAvailableParkingSpots

If obstacles in parking_spots we are deleting that nodes from current parking spots

IsObstacleStillThere

In next loop, checking knowed obstacles again for being sure is the obstacle still there or just moved away. Also this scenerio will work(soon) for if the path is not dashed line.